fix: 净值趋势卡从首页迁到资产页(与App端位置一致)
